+++ dom/bootstrap.js
@@ -1,6 +1,8 @@
 /* eslint-env browser */
 
 import { RNDomInstance } from 'react-native-dom';
+import FontLoader from 'react-native-dom-expo/native/font-loader';
+
 import appJson from '../app.json';
 
 // Path to RN Bundle Entrypoint ===============================
@@ -9,7 +11,7 @@ const rnBundlePath = './entry.bundle?platform=dom&dev=true';
 // React Native DOM Runtime Options ===========================
 const ReactNativeDomOptions = {
   enableHotReload: false,
-  nativeModules: [],
+  nativeModules: [FontLoader],
 };
 
 // App Initialization =========================================