+++ src/ExpoFontLoader.dom.js
@@ -1,9 +1,16 @@
+import Constants from 'expo-constants';
+
+import { NativeModules } from 'react-native';
+
+const { FontLoader } = NativeModules;
+
 export default {
   get name() {
     return 'ExpoFontLoader';
   },
-  loadAsync(fontFamilyName, resource) {
-    console.log('font loader not impelmented', fontFamilyName, resource);
-    throw new Error('font loader not implemented');
+  async loadAsync(nativeFontName, resource) {
+    // To revert `${Constants.sessionId}-${name}`
+    const fontName = nativeFontName.slice(Constants.sessionId.length + 1);
+    return FontLoader.loadFont(fontName, resource);
   },
 };